refactor(api): move projects, sessions, metrics dashboard endpoints
Commit 3 of the dashboard fold-in. Lifts 8 endpoints out of
dashboard/api/main.py into three new files under
server/api/dashboard/:

- projects.py: GET /api/projects, GET+POST /api/projects/{id}/guardrails
- sessions.py: GET /api/sessions, GET /api/sessions/{id},
               GET /api/sessions/{id}/turns, /api/sessions/{id}/dead_air
- metrics.py:  GET /api/metrics

Each handler swaps the module-global ._gateway lookup for
Depends(get_gateway). Behaviour is preserved verbatim: the metrics
endpoint keeps its raw text() SQL + dead-air id-join, the guardrails
POST keeps its Pydantic ValidationError -> 400 conversion path, the
sessions list keeps the order_by regex pattern.

tests/server/test_metrics_endpoint.py and the one
dash_module.app test in test_dashboard_api.py both repoint at
the daemon's build_app(...). Without it, the deleted handlers on
the legacy dashboard app would get caught by the SPA fallback and
return the index HTML instead of a 200 with the expected JSON shape.

ruff check src + uv run mypy + uv run pytest -q all green:
1666 passed, 4 skipped.
